Require loading as a module
The issue is wgsl_reflect defines `class Function`. If this code
is loaded as a non-module then that ends up overriding the default
`Function` class and code that's checking `... instanceof Function`
starts failing.

There's no easy way for a script to check if it was loaded as a
module or a non-module but using `import.meta` ends up requiring
it to be loaded as a module.